This does everything to prepare an ubuntu server to run my code.
This is to make it a really smooth process for anyone who wants to use the code. I plan to make one for fedora/RedHat as well, but for now I am just making one for ubuntu. So once this is run it can run any of my code flawlessly.