Commented and deleted RSA signing
The reason signing got deleted is because to sign the files after encrypting them takes far too long, requires two sets of public and  prvate keys and I think I was doing it wrong entirely because I was unable to get the contents of the file out of the signed files. The Golang libraries for RSA do not return anything other than whether or not it verrified. I am not able to get the contents of the file so I am going to stick to AES and transmiting that with RSA.