Streamlining by using bootstrap min instead of bootsrap
This cuts down majorly on the size of the code and the number of lines we have to work with. It takes all the bootsrap and puts it on one line that really takes up about 50 lines because it iss so long, but it avoids the 7000 lines that it used to be and also brings down the size of the file by 25 kb which is great for low bandwidth areas.